refactor($기타): useEffect를 사용하여 로그인 여부를 확인하는 반복되는 로직을 훅으로 분리 (/useAuth)
useEffect를 사용하여 토큰을 확인하는 로직이 반복되어, 이를 줄이고자 훅으로 분리
