+++ FE/src/components/common/header/HeaderNavSection.tsx
@@ -1,24 +1,17 @@
 "use client";
-import { useEffect, useState } from "react";
+
 import CreateBtn from "./CreateBtn";
 import LoginRedirectBtn from "./LoginRedirectBtn";
 import ManageRedirectButton from "./ManageRedirectButton";
 import UserBtn from "./UserBtn";
-import { CheckIsLoggedIn } from "@/utils/authWithStorage";
+import useAuth from "@/hooks/useAuth";
 
 interface HeaderNavSectionProps {
   isAdmin: boolean;
 }
 
 const HeaderNavSection = ({ isAdmin }: HeaderNavSectionProps) => {
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
-  const [isLoading, setIsLoading] = useState(true);
-
-  useEffect(() => {
-    const isLoggedIn = CheckIsLoggedIn();
-    setIsLoggedIn(isLoggedIn);
-    setIsLoading(false);
-  }, []);
+  const { isLoading, isLoggedIn } = useAuth();
 
   if (isLoading) return null;
   if (isAdmin)
